requirements.txt
ultralytics
opencv-python
numpy
mediapipe
scipy
python-rtmidi
python-osc
